+++ Testing/group_test.py
@@ -1,6 +1,5 @@
 from player_generator import new_dict
-from mythic_groups_maker import *
-# from sqlconnector.sqlReader import sqlPlayerDict
+from mythicgroupmaker.mythic_groups_maker import *
 
 sqlPlayerDict = new_dict
 print("Importing dictionary ... ...")
