+++ Testing/player_generator.py
