+++ sqlconnector/connect_tcp.py
@@ -1,21 +1,3 @@
-# Copyright 2022 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# TODO (https://github.com/GoogleCloudPlatform/python-docs-samples/issues/8253): remove old region tags
-# [START cloud_sql_mysql_sqlalchemy_connect_tcp]
-# [START cloud_sql_mysql_sqlalchemy_sslcerts]
-# [START cloud_sql_mysql_sqlalchemy_connect_tcp_sslcerts]
 import os
 
 import sqlalchemy
